Update README.md
# HELP-
To the Public!
My name is ANGELICA MAYER-TUCKER
THE REAL ANGELICA MAYER TUCKER. THE OWNER OF rangelica.8020@outlook.com

To the public, this grit hub site has been attached to my name and for security reasons I am not allowed to close it down nor have access to my Outlook profile. My guess is because it's being investigated I don't blame Microsoft because this site created is FAKE I NEVER CREATED IT. My email account HAS been high jacked for some time now but I cannot let not having access to my own account go on any longer. 
In this grit hub it says I have an owner of this "supposed" company for 5 years but I've only had the email address for less than a year. I am asking any and all who read this for help. I am being cyber bullied and need help properly reporting this because I can not use my phone or tablet which is what I am on now because it's infected. I don't have control over this. So I could post and as fast as I post who ever controlling my account will take it down. I need help. I am going to make a face book account asking specifically for the help. 
PLEASE! TO THE PUBLIC, 
I AM LETTING ANY PERSON, BUSINESS, ENTITY know THAT THIS IS A FAKE COMPANY IT DOES NOT EXIST AND WAS NOT CREATED BY ME ANGELICA MAYER-TUCKER. 
I WILL HAVE A CURRENT PHONE NUMBER THAT BELONGS TO ME SOON SO THAT I CAN ADD IT TO THIS REPOSITORY. 
WILL UPDATE SOON. FOR NOW PLEASE DO NOT USE THIS AS ANY KIND OF REFERENCE TO MY NAME. I DO NOT WORK AT MRA-JAC CHICKEN WHATEVER THAT IT'S BUT AM ALSO NOT A CHICKEN FOR POSTING THIS! 
THANK YOU FOR YOUR TIME PUBLIC.
ANGELICA MAYER-TUCKER